+++ example/lib/main.dart
@@ -7,9 +7,11 @@ import 'routes.dart';
 
 void main() {
   Get.put(Database(), permanent: true);
+  QR.enableDebugLog = true;
   runApp(MyApp());
 }
 
+/// See AppRoutes definition https://github.com/SchabanBo/qlevar_router/blob/master/example/lib/routes.dart
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) => MaterialApp.router(