+++ lib/src/types.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
+
+import 'qr.dart';
 import 'routes_tree.dart';
 
+/// QRouter to palce where the children of natsten route should appear.
 class QRouter<T> extends Router<T> {
   const QRouter({
     Key key,
@@ -23,6 +26,8 @@ typedef RedirectGuard = String Function(String);
 /// [name] the name of the route.
 /// [path] the path of the route.
 /// [page] the page to show
+/// [onInit] a function to do what you need before initializing the route.
+/// [onDispose] a function to do what you need before disposing the route.
 /// It give the child router to use it in the parent page
 /// when the route has children, otherwise it give null.
 /// [redirectGuard] it gives the called path and takes the new path
@@ -33,24 +38,40 @@ class QRoute {
   final String path;
   final QRouteBuilder page;
   final RedirectGuard redirectGuard;
+  final Function onInit;
+  final Function onDispose;
   final List<QRoute> children;
 
   QRoute(
       {this.name,
       @required this.path,
-      this.redirectGuard,
       this.page,
+      this.onInit,
+      this.onDispose,
+      this.redirectGuard,
       this.children});
+
+  QRoute copyWith({
+    String name,
+    String path,
+    QRouteBuilder page,
+    RedirectGuard redirectGuard,
+    Function onInit,
+    Function onDispose,
+    List<QRoute> children,
+  }) =>
+      QRoute(
+        name: name ?? this.name,
+        path: path ?? this.path,
+        page: page ?? this.page,
+        redirectGuard: redirectGuard ?? this.redirectGuard,
+        onInit: onInit ?? this.onInit,
+        onDispose: onDispose ?? this.onDispose,
+        children: children ?? this.children,
+      );
 }
 
-/// The cureent route inforamtion 
-class QCurrentRoute {
-  String fullPath = '';
-  Map<String, dynamic> params = {};
-  MatchContext match;
-}
-
-/// the match context for a route
+/// The match context for a route.
 class MatchContext {
   final int key;
   final String name;
@@ -58,7 +79,10 @@ class MatchContext {
   String cureentPath;
   final bool isComponent;
   final QRouteBuilder page;
+  final Function onInit;
+  final Function onDispose;
   MatchContext childContext;
+  QNavigationMode mode;
   QRouter<dynamic> router;
 
   MatchContext(
@@ -66,7 +90,10 @@ class MatchContext {
       this.key,
       this.fullPath,
       this.isComponent,
+      this.onInit,
+      this.onDispose,
       this.page,
+      this.mode,
       this.childContext,
       this.router});
 
@@ -74,6 +101,8 @@ class MatchContext {
       key: key,
       name: name,
       fullPath: fullPath ?? this.fullPath,
+      onInit: onInit,
+      onDispose: onDispose,
       isComponent: isComponent ?? this.isComponent,
       page: page,
       childContext: childContext,
@@ -84,6 +113,8 @@ class MatchContext {
       MatchContext(
           name: route.route.name,
           isComponent: route.route.isComponent,
+          onDispose: route.route.onDispose,
+          onInit: route.route.onInit,
           key: route.route.key,
           fullPath: route.route.fullPath,
           page: route.route.page,
@@ -93,10 +124,17 @@ class MatchContext {
   MaterialPage toMaterialPage() =>
       MaterialPage(name: name, key: ValueKey(fullPath), child: page(router));
 
-  void updatePath(String path){
+  void updatePath(String path) {
     cureentPath = path;
-    if (childContext!=null) {
+    if (childContext != null) {
       childContext.updatePath(path);
+    }
+  }
+
+  void setNavigationMode(QNavigationMode nMode) {
+    mode = nMode;
+    if (childContext != null) {
+      childContext.setNavigationMode(nMode);
     }
   }
 
@@ -107,8 +145,6 @@ class MatchContext {
     router.routerDelegate.setNewRoutePath(childContext);
   }
 
-  bool isMatch(MatchContext other) => other.isComponent
-      ? fullPath == other.fullPath
-      : key == other.key;
-
+  bool isMatch(MatchContext other) =>
+      other.isComponent ? fullPath == other.fullPath : key == other.key;
 }